pyproject.toml
[project]
name = "hifi-gan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.11.0",
    "comet-ml>=3.54.2",
    "hydra-core>=1.3.2",
    "isort>=7.0.0",
    "librosa>=0.11.0",
    "matplotlib>=3.10.7",
    "numpy==1.26.4",
    "pandas>=2.3.3",
    "pre-commit>=4.5.0",
    "ruff>=0.14.8",
    "speechbrain>=1.0.3",
    "torch==2.7.0+cu128; sys_platform != 'darwin'",
    "torch>=2.7.0; sys_platform == 'darwin' and platform_machine == 'arm64'",
    "torchaudio==2.7.0+cu128; sys_platform != 'darwin'",
    "torchaudio>=2.7.0; sys_platform == 'darwin' and platform_machine == 'arm64'",
    "torchvision==0.22.0+cu128; sys_platform != 'darwin'",
    "torchvision>=0.22.0; sys_platform == 'darwin' and platform_machine == 'arm64'",
    "torchmetrics>=1.8.2",
    "tqdm>=4.67.1",
    "wandb>=0.23.1",
    "wvmos",
    "einops>=0.8.1",
    "gdown>=5.2.0",
]

[tool.uv]
package = true
python-preference = "managed"

[tool.uv.sources]
torch = [
    { index = "pytorch", marker = "sys_platform != 'darwin'" },
    { index = "default", marker = "sys_platform == 'darwin'" }
]
torchvision = [
    { index = "pytorch", marker = "sys_platform != 'darwin'" },
    { index = "default", marker = "sys_platform == 'darwin'" }
]
torchaudio = [
    { index = "pytorch", marker = "sys_platform != 'darwin'" },
    { index = "default", marker = "sys_platform == 'darwin'" }
]
wvmos = { git = "https://github.com/AndreevP/wvmos" }

[[tool.uv.index]]
name = "default"
url = "https://pypi.org/simple"

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu128"


[tool.pytest.ini_options]
log_cli = true
log_cli_level = "WARNING"
addopts = "-s"

[tool.ruff]
line-length = 120

[tool.ruff.lint.extend-per-file-ignores]
# Also ignore `E402` in all `__init__.py` files.
"__init__.py" = ["E402", "F401", "F403", "E501"]

[tool.ruff.lint]
extend-select = ["UP"]
ignore = [
    "F722"
]
